Update issue 1198 Don't drop performance_schema table.
git-svn-id: http://maatkit.googlecode.com/svn/trunk@7140 dfb901c2-3250-0410-b216-0b33211c9131

+++ common/Sandbox.pm
@@ -152,6 +152,7 @@ sub wipe_clean {
    foreach my $db ( @{$dbh->selectcol_arrayref('SHOW DATABASES')} ) {
       next if $db eq 'mysql';
       next if $db eq 'information_schema';
+      next if $db eq 'performance_schema';
       next if $db eq 'sakila';
       $dbh->do("DROP DATABASE IF EXISTS `$db`");
    }
